added: BBS: "Connection Script". The route to the FWD partner is established step by step by sending commands for the respective connection setup.
E.g.: DX0SAW - c DX0SAW
DX0BBS - m
The next step is identified by the respective flag "*** Connected to <CALL>".

Alternatively, a timer can be set to specify how long to wait until the
next step (sending the command). If a timer is set, the "*** Connected to <CALL>" flag is ignored.